Add a module to fn for direct conversion between Julian and Gregorian.
There are utility methods. We hope the organization of the package means it is useful to create a date in one calendar, then convert it to a date in another calendar and examine the result. But we include these functions for the case where you want to find the corresponding day, month and year for a Julian and a Gregorian date in a straightforward way without creating any objects explicitly.