Add pragma: no cover.
This line never gets used as we always check for a julian leap year first and we never hit this branch when we have already confirmed that its a julian leap year. However it seems unnecesary to add a test for the private class, and unelegant to make a version of this function which doesnt work for any year.